Remove explicit support for Windows 32-bit builds
Windows 32-bit builds are rarely necessary, and removing the explicit
handling of 32-bit vs 64-bit settings simplifies setup and configuration
on Windows. For example, only one `assets_os/windows` needs to be kept
for Windows assets, reducing duplication. The Windows-specific flags
(lines 52-59) are also simpler now.

The compiled binaries will be 64-bit if compiled using a 64-bit version
of MinGW-w64's GCC, and 32-bit if compiled using a 32-bit version of
MinGW-w64's GCC. This is the default behavior, and adding -m32 or -m64
did not work in any case when used with the wrong compiler (e.g. passing
-m64 to a 32-bit GCC).