Remove install target
The provided install target was not compliant with expectations for
`make install`. This should respect conventions like the install prefix,
splitting files into bin, lib, and share, and also platform-specific
conventions. For the sake of simplicity, it has been removed for now, as
its removal is preferable to its presence in a misleading state.