Move bin/ directory under build/ directory
Having separate top-level build/ and bin/ directories was an artifact of
Code::Blocks's file structure, which had served as inspiration. However,
this is not a very common convention for other C++ build systems and in
other languages.

This changes the directory for the binary to be under
build/<platform>/<configuration>/bin. The object files (as well as
dependency files and compilation database fragments) are now in
build/<platform>/<configuration>/obj to match.

This will also be helpful for when several binaries, or even libraries,
are produced, preventing the need for yet another top-level directory.