Disable default implicit rules
This avoids acccidentally compiling files outside of the source tree
through the built-in implicit rules. For example, if a.cpp was at the
root of the project, a.o could be made using `make a.o`. This is an
unintended workflow enabled by the default implicit rules. The default
implicit rules which are unneeded because the Makefile has its own
pattern rules for everything which is needed.