Make separate include directory optional
This also adds the src and tests directories to the preprocessor's
search path by default, as it should be possible to use these
directories for private headers even when public headers are placed in
the include directory.

Doing so also now allows projects to forego using a separate include
directory, and place all their header files under src and tests.